requirements.txt
# Dependências principais do Jarvis
eel==0.18.2
pyttsx3==2.90
SpeechRecognition==3.10.0
opencv-python==4.8.1.78
pyaudio==0.2.11
pywhatkit==5.4
pvporcupine==3.0.2
requests==2.31.0
playsound==1.3.0
pyautogui==0.9.54
google-generativeai==0.3.2
face-recognition==1.3.0
dlib==19.24.2
python-dotenv==1.0.0

# Dependências web
Flask==2.0.1
firebase-admin==5.2.0

# Dependências de áudio e vídeo
Pillow==10.0.1
numpy==1.24.3

# Dependências opcionais para funcionalidades avançadas
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
gevent==23.9.1
gevent-websocket==0.10.1
hugchat==0.4.0